Unite prediction market parameters
This commit also restructures the project, adding more granularity to the primitives crate
