+++ node/src/chain_spec/mod.rs
@@ -7,7 +7,7 @@ pub use dev::dev_config;
 pub use local_testnet::local_testnet_config;
 use sp_core::{Pair, Public};
 use sp_runtime::traits::{IdentifyAccount, Verify};
-use zeitgeist_primitives::{AccountId, Signature};
+use zeitgeist_primitives::types::{AccountId, Signature};
 use zeitgeist_runtime::{BalancesConfig, TokensConfig};
 
 const TELEMETRY_URL: &str = "wss://telemetry.polkadot.io/submit/";
