+++ primitives/src/traits.rs
@@ -1,0 +1,3 @@
+pub use crate::swaps::Swaps;
+pub use crate::zeitgeist_currencies_extension::ZeitgeistCurrenciesExtension;
+pub use crate::zeitgeist_multi_reservable_currency::ZeitgeistMultiReservableCurrency;
